pyproject.toml
[tool.poetry]
name = "faceblur"
version = "0.1.0"
description = ""
authors = ["Sebastian Koch "]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
pillow = "^10.0.0"
numpy = "^1.25.1"
opencv-python-headless = "^4.8.0.74"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
